Make colcade calculate image sizes correctly
Before, all the grid-cols and grid-items were getting place next to each 
other in the grids flex box. This made it so that the width of the 
columns was all off, since the 4 columns wouldn't actually span the 
entire wide. This made it so images were getting placed in the wrong 
place.

Additionally, since colcade might run before the images are loaded, the 
height of each item is garbage. To fix this, add a placeholder div that 
it underneath the image to begin with, so the size is calculated 
correctly, and once the image loads it is covered up.